refactor: rename semantics_base.py to registry_base.py
