+++ docs/reference/api/registry_base.md
@@ -1,11 +1,11 @@
-# `raitap.semantics_base`
+# `raitap.registry_base`
 
 Cross-module interface declared by every adapter that publishes an
 algorithm-name → typed-hints registry. Used by both the transparency
 explainer adapters and the robustness assessor adapters.
 
 ```{eval-rst}
-.. autoclass:: raitap.semantics_base.WithAlgorithmRegistry
+.. autoclass:: raitap.registry_base.WithAlgorithmRegistry
    :members:
    :show-inheritance:
 ```
