+++ src/raitap/robustness/assessors/base_assessor.py
@@ -24,7 +24,7 @@ import torch
 
 from raitap import raitap_log
 from raitap.configs import resolve_run_dir
-from raitap.semantics_base import WithAlgorithmRegistry
+from raitap.registry_base import WithAlgorithmRegistry
 
 from ..contracts import (
     MethodKind,
@@ -63,7 +63,7 @@ class BaseAssessor(WithAlgorithmRegistry["AssessorSemanticsHints"], abstract=Tru
 
     Concrete subclasses must declare ``algorithm_registry: ClassVar[Mapping[str,
     AssessorSemanticsHints]]`` per the
-    :class:`raitap.semantics_base.WithAlgorithmRegistry` contract.
+    :class:`raitap.registry_base.WithAlgorithmRegistry` contract.
     Intermediate abstract classes opt out via ``abstract=True``.
     """
